requirements.txt
pyperclip
python-docx
pytest
